requirements.txt
Flask-SQLAlchemy
Flask-Migrate
psycopg2
PyJWT